Add files via upload
Used GPT-4o to port the original poe to use Gemini 2.5 Pro on the GCP Vertex framework. Complete feature parity in 1.0, but in 1.1 , swapped out for Vertex built in grounding so you don't need serpapi anymore. In 1.2 I put inline blocking using model guard with a template that calls sub process to gcloud to set it up. You will need the appropriate IAM permissions to run. This assumes you have an active session in Cloudshell or other post-auth GCP session.